# Make PageProcessor preserve input block laziness