--- a/presto-main/src/main/java/io/prestosql/operator/ScanFilterAndProjectOperator.java
+++ b/presto-main/src/main/java/io/prestosql/operator/ScanFilterAndProjectOperator.java
@@ -424,6 +424,12 @@ public class ScanFilterAndProjectOperator
         }
 
         @Override
+        public PlanNodeId getPlanNodeId()
+        {
+            return planNodeId;
+        }
+
+        @Override
         public String getOperatorType()
         {
             return ScanFilterAndProjectOperator.class.getSimpleName();