--- a/presto-main/src/main/java/io/prestosql/operator/ScanFilterAndProjectOperator.java
+++ b/presto-main/src/main/java/io/prestosql/operator/ScanFilterAndProjectOperator.java
@@ -84,7 +84,8 @@ public class ScanFilterAndProjectOperator
             Iterable<ColumnHandle> columns,
             Iterable<Type> types,
             DataSize minOutputPageSize,
-            int minOutputPageRowCount)
+            int minOutputPageRowCount,
+            boolean avoidPageMaterialization)
     {
         pages = splits.flatTransform(
                 new SplitToPages(
@@ -98,7 +99,8 @@ public class ScanFilterAndProjectOperator
                         types,
                         requireNonNull(memoryTrackingContext, "memoryTrackingContext is null").aggregateSystemMemoryContext(),
                         minOutputPageSize,
-                        minOutputPageRowCount));
+                        minOutputPageRowCount,
+                        avoidPageMaterialization));
     }
 
     @Override
@@ -181,6 +183,7 @@ public class ScanFilterAndProjectOperator
         final LocalMemoryContext outputMemoryContext;
         final DataSize minOutputPageSize;
         final int minOutputPageRowCount;
+        final boolean avoidPageMaterialization;
 
         SplitToPages(
                 Session session,
@@ -193,7 +196,8 @@ public class ScanFilterAndProjectOperator
                 Iterable<Type> types,
                 AggregatedMemoryContext aggregatedMemoryContext,
                 DataSize minOutputPageSize,
-                int minOutputPageRowCount)
+                int minOutputPageRowCount,
+                boolean avoidPageMaterialization)
         {
             this.session = requireNonNull(session, "session is null");
             this.yieldSignal = requireNonNull(yieldSignal, "yieldSignal is null");
@@ -209,6 +213,7 @@ public class ScanFilterAndProjectOperator
             this.outputMemoryContext = localAggregatedMemoryContext.newLocalMemoryContext(ScanFilterAndProjectOperator.class.getSimpleName());
             this.minOutputPageSize = requireNonNull(minOutputPageSize, "minOutputPageSize is null");
             this.minOutputPageRowCount = minOutputPageRowCount;
+            this.avoidPageMaterialization = avoidPageMaterialization;
         }
 
         @Override
@@ -256,7 +261,8 @@ public class ScanFilterAndProjectOperator
                             session.toConnectorSession(),
                             yieldSignal,
                             outputMemoryContext,
-                            page))
+                            page,
+                            avoidPageMaterialization))
                     .transformProcessor(processor -> mergePages(types, minOutputPageSize.toBytes(), minOutputPageRowCount, processor, localAggregatedMemoryContext))
                     .withProcessStateMonitor(state -> memoryContext.setBytes(localAggregatedMemoryContext.getBytes()));
         }
@@ -451,6 +457,24 @@ public class ScanFilterAndProjectOperator
                 DriverYieldSignal yieldSignal,
                 WorkProcessor<Split> splits)
         {
+            return create(session, memoryTrackingContext, yieldSignal, splits, true);
+        }
+
+        public WorkProcessorSourceOperator createAdapterOperator(Session session,
+                MemoryTrackingContext memoryTrackingContext,
+                DriverYieldSignal yieldSignal,
+                WorkProcessor<Split> splits)
+        {
+            return create(session, memoryTrackingContext, yieldSignal, splits, false);
+        }
+
+        private ScanFilterAndProjectOperator create(
+                Session session,
+                MemoryTrackingContext memoryTrackingContext,
+                DriverYieldSignal yieldSignal,
+                WorkProcessor<Split> splits,
+                boolean avoidPageMaterialization)
+        {
             return new ScanFilterAndProjectOperator(
                     session,
                     memoryTrackingContext,
@@ -463,7 +487,8 @@ public class ScanFilterAndProjectOperator
                     columns,
                     types,
                     minOutputPageSize,
-                    minOutputPageRowCount);
+                    minOutputPageRowCount,
+                    avoidPageMaterialization);
         }
 
         @Override