--- a/presto-main/src/main/java/io/prestosql/operator/WorkProcessorSourceOperatorFactory.java
+++ b/presto-main/src/main/java/io/prestosql/operator/WorkProcessorSourceOperatorFactory.java
@@ -24,6 +24,8 @@ public interface WorkProcessorSourceOperatorFactory
 
     PlanNodeId getSourceId();
 
+    PlanNodeId getPlanNodeId();
+
     String getOperatorType();
 
     WorkProcessorSourceOperator create(