--- a/presto-main/src/test/java/io/prestosql/operator/TestWorkProcessorPipelineSourceOperator.java
+++ b/presto-main/src/test/java/io/prestosql/operator/TestWorkProcessorPipelineSourceOperator.java
@@ -248,6 +248,12 @@ public class TestWorkProcessorPipelineSourceOperator
         }
 
         @Override
+        public PlanNodeId getPlanNodeId()
+        {
+            return sourceId;
+        }
+
+        @Override
         public String getOperatorType()
         {
             return TestWorkProcessorSourceOperatorFactory.class.getSimpleName();