… presto-main/src/test/java/io/prestosql/operator/project/TestDictionaryAwarePageProjection.java
… presto-main/src/test/java/io/prestosql/operator/project/TestPageProcessor.java