Followup:
* port semi-join to `WorkProcessorOperator` (@Praveen2112 ?)
* port `FilterAndProjectOperator` to `WorkProcessorOperator`
This way we get (node-local) dynamic filtering for semi joins and `IN (subquery), NOT IN (subquery), x >= ALL (subquery), etc` expressions