@sopel39 2019-07-08
Benchmarks:
```
AFTER:

Benchmark                                                      (columnCount)  (dictionaryBlocks)  (positionsPerPage)   (type)  Mode  Cnt    Score   Error  Units
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                  32  varchar  avgt   50  141.904 ± 0.821  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                  32   bigint  avgt   50   45.060 ± 0.567  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                1024  varchar  avgt   50   77.402 ± 0.540  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                1024   bigint  avgt   50    9.260 ± 0.039  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                  32  varchar  avgt   50   90.962 ± 0.611  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                  32   bigint  avgt   50   56.806 ± 0.832  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                1024  varchar  avgt   50   19.393 ± 0.171  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                1024   bigint  avgt   50    8.482 ± 0.017  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                  32  varchar  avgt   50  211.098 ± 0.377  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                  32   bigint  avgt   50   72.684 ± 0.713  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                1024  varchar  avgt   50   96.418 ± 0.502  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                1024   bigint  avgt   50   13.353 ± 0.073  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                  32  varchar  avgt   50  149.502 ± 0.675  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                  32   bigint  avgt   50   94.367 ± 1.635  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                1024  varchar  avgt   50   24.067 ± 0.242  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                1024   bigint  avgt   50   12.741 ± 0.066  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                  32   bigint  avgt   50  125.515 ± 1.271  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                1024  varchar  avgt   50  130.306 ± 0.741  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                1024   bigint  avgt   50   21.516 ± 0.048  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                  32  varchar  avgt   50  256.790 ± 1.033  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                  32   bigint  avgt   50  162.861 ± 2.340  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                1024  varchar  avgt   50   33.036 ± 0.096  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                1024   bigint  avgt   50   21.007 ± 0.081  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                  32   bigint  avgt   50  224.977 ± 1.399  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                1024  varchar  avgt   50  200.095 ± 0.717  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                1024   bigint  avgt   50   37.257 ± 0.113  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                  32  varchar  avgt   50  484.378 ± 1.285  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                  32   bigint  avgt   50  292.979 ± 1.243  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                1024  varchar  avgt   50   50.663 ± 0.503  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                1024   bigint  avgt   50   37.043 ± 0.152  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                  32   bigint  avgt   50  420.169 ± 2.380  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                1024  varchar  avgt   50  342.290 ± 1.434  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                1024   bigint  avgt   50   69.130 ± 0.343  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                  32   bigint  avgt   50  612.408 ± 4.062  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                1024  varchar  avgt   50   87.303 ± 0.618  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                1024   bigint  avgt   50   71.148 ± 0.465  ms/op

BEFORE:

Benchmark                                                      (columnCount)  (dictionaryBlocks)  (positionsPerPage)   (type)  Mode  Cnt    Score    Error  Units
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                  32  varchar  avgt   50  143.775 ±  1.108  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                  32   bigint  avgt   50   45.893 ±  0.651  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                1024  varchar  avgt   50   77.121 ±  0.609  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2               false                1024   bigint  avgt   50    9.616 ±  0.051  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                  32  varchar  avgt   50   92.151 ±  0.384  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                  32   bigint  avgt   50   56.643 ±  0.310  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                1024  varchar  avgt   50   19.985 ±  0.174  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              2                true                1024   bigint  avgt   50    8.833 ±  0.043  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                  32  varchar  avgt   50  214.515 ±  0.780  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                  32   bigint  avgt   50   75.181 ±  1.003  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                1024  varchar  avgt   50   95.147 ±  0.377  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4               false                1024   bigint  avgt   50   13.822 ±  0.051  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                  32  varchar  avgt   50  149.661 ±  0.613  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                  32   bigint  avgt   50   94.085 ±  0.839  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                1024  varchar  avgt   50   24.553 ±  0.174  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              4                true                1024   bigint  avgt   50   13.018 ±  0.051  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                  32   bigint  avgt   50  127.073 ±  1.259  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                1024  varchar  avgt   50  129.735 ±  0.541  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8               false                1024   bigint  avgt   50   21.865 ±  0.104  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                  32  varchar  avgt   50  258.460 ±  0.951  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                  32   bigint  avgt   50  165.347 ±  2.190  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                1024  varchar  avgt   50   33.486 ±  0.182  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented              8                true                1024   bigint  avgt   50   21.073 ±  0.075  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                  32   bigint  avgt   50  221.195 ±  1.120  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                1024  varchar  avgt   50  199.427 ±  0.476  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16               false                1024   bigint  avgt   50   37.843 ±  0.175  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                  32  varchar  avgt   50  485.161 ±  1.569  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                  32   bigint  avgt   50  295.203 ±  2.633  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                1024  varchar  avgt   50   52.209 ±  0.265  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             16                true                1024   bigint  avgt   50   37.172 ±  0.220  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                  32   bigint  avgt   50  413.407 ±  1.834  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                1024  varchar  avgt   50  339.468 ±  1.265  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32               false                1024   bigint  avgt   50   69.612 ±  0.218  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                  32   bigint  avgt   50  584.004 ± 12.342  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                1024  varchar  avgt   50   89.021 ±  0.305  ms/op
BenchmarkScanFilterAndProjectOperator.benchmarkColumnOriented             32                true                1024   bigint  avgt   50   71.113 ±  0.421  ms/op
```